+++ DatadogCore/Sources/Core/Upload/URLSessionClient.swift
@@ -39,9 +39,7 @@ internal class URLSessionClient: HTTPClient {
         let task = session.dataTask(with: request) { data, response, error in
             completion(httpClientResult(for: (data, response, error)))
         }
-        if #available(iOS 15.0, tvOS 15.0, watchOS 8.0, *) {
-            task.delegate = delegate
-        }
+        task.delegate = delegate
         task.resume()
     }
 }